UserAuth/requirements.txt
Django==4.2.3
requests
psycopg2>=2.8
djangorestframework==3.14.0
django-rest-knox==4.2.0
django-cors-headers==4.2.0
django-filter==23.2
pika

pytest-cov==4.1.0
pytest-django==4.5.2
mixer==7.2.2